Remove cyclic dependencies surrounding `H5Easy.hpp`.
